requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-telegram-bot==20.7
schedule==1.2.0
python-dotenv==1.0.0